requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --strip-extras requirement.in
#
annotated-types==0.6.0
    # via
    #   -r requirement.in
    #   pydantic
anyio==3.7.1
    # via
    #   -r requirement.in
    #   fastapi
    #   starlette
click==8.1.7
    # via
    #   -r requirement.in
    #   uvicorn
dnspython==2.4.2
    # via
    #   -r requirement.in
    #   email-validator
    #   pymongo
email-validator==2.1.0.post1
    # via -r requirement.in
fastapi==0.104.1
    # via -r requirement.in
h11==0.14.0
    # via
    #   -r requirement.in
    #   uvicorn
idna==3.4
    # via
    #   -r requirement.in
    #   anyio
    #   email-validator
motor==3.3.1
    # via -r requirement.in
pydantic~=2.6.4
    # via
    #   -r requirement.in
    #   fastapi
pydantic-core==2.10.1
    # via
    #   -r requirement.in
    #   pydantic
pymongo==4.5.0
    # via
    #   -r requirement.in
    #   motor
python-dotenv==1.0.1
    # via -r requirement.in
sniffio==1.3.0
    # via
    #   -r requirement.in
    #   anyio
starlette==0.27.0
    # via
    #   -r requirement.in
    #   fastapi
typing-extensions==4.8.0
    # via
    #   -r requirement.in
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.23.2
    # via -r requirement.in

beanie~=1.25.0
passlib~=1.7.4